+++ cypress/support/e2e.js
@@ -280,5 +280,45 @@ Cypress.Commands.add('c_emailVerificationSignUp', (verification_code,event_email
             }
           })
   })
-    
+  
+const pixelmatch = require('pixelmatch')
+const { PNG } = require('pngjs')
+
+Cypress.Commands.add('compareElementScreenshots', (elementSelector, imageName1, imageName2, diffImageName) => {
+  const timestamp = new Date().getTime()
+  
+  // To hide the timestamp in the footer
+  cy.get('[class="dc-popover server-time"]').invoke('css', 'visibility', 'hidden')
+  
+  cy.get(elementSelector).then(($el) => {
+    const { top, left, width, height } = $el[0].getBoundingClientRect()
+
+    cy.screenshot(`${imageName1}_${timestamp}`, { clip: { x: left, y: top, width, height } })
+
+    cy.wait(2000)  // 2 seconds wait time for new tick
+
+    cy.screenshot(`${imageName2}_${timestamp}`, { clip: { x: left, y: top, width, height } })
+  })
+
+  cy.readFile(`./cypress/screenshots/${imageName1}_${timestamp}.png`, 'base64').then((img1Data) => {
+    cy.readFile(`./cypress/screenshots/${imageName2}_${timestamp}.png`, 'base64').then((img2Data) => {
+      const img1 = PNG.sync.read(Buffer.from(img1Data, 'base64'))
+      const img2 = PNG.sync.read(Buffer.from(img2Data, 'base64'))
+      const { width, height } = img1
+      const diff = new PNG({ width, height })
+
+      
+      const mismatchedPixels = pixelmatch(img1.data, img2.data, diff.data, width, height, {
+        threshold: 0.1, // this threshold is to adjust the sensitivity of the mismatched pixels
+      })
+      
+      cy.writeFile(`./cypress/screenshots/${diffImageName}_${timestamp}.png`, PNG.sync.write(diff), 'base64')
+      
+      expect(mismatchedPixels).to.be.greaterThan(0) // we expect the feed is updating, so the mismatchedPixels must be more than 0 to prove differences in both screenshots
+
+    })
+
+  })
+})
+
 })
